Add Getter and Update Accessibility
The language model's name cannot be accessed outside the model. Now
add a getter to get the model's name.

Methods tokenize(), add_tokens() are not used outside the model. Now
changed them to be private methods.

The function of the method add_string() is more acurately described as
add string to train the model. Now its name is changed to be train()

Sort methods according to accessibility.